Validations on null requests to `MarcaController` routes
Use `response` helper with json

+++ 01_Laravel_VueJs/04_locadora_api-rest/app/Http/Controllers/MarcaController.php
@@ -17,34 +17,47 @@ class MarcaController extends Controller
     public function index()
     {
         $marcas = $this->marca->all();
-        return $marcas;
+        return response()->json($marcas, 200);
     }
 
     public function store(Request $request)
     {
         $marca = $this->marca->create($request->all());
-        return $marca;
+        return response()->json($marca, 201);
     }
 
     public function show(Int $id)
     {
         $marca = $this->marca->find($id);
-        return $marca;
+
+        if (!$marca) {
+            return response()->json(['erro' => 'Marca não encontrada.'], 404);
+        }
+
+        return response()->json($marca, 200);
     }
 
     public function update(Request $request, Int $id)
     {
         $marca = $this->marca->find($id);
-        $marca->update($request->all());
 
-        return $marca;
+        if (!$marca) {
+            return response()->json(['erro' => 'Marca não encontrada.'], 404);
+        }
+
+        $marca->update($request->all());
+        return response()->json($marca, 200);
     }
 
     public function destroy(Int $id)
     {
         $marca = $this->marca->find($id);
-        $marca->delete();
 
-        return ['msg' => 'A marca foi removida com sucesso!'];
+        if (!$marca) {
+            return response()->json(['erro' => 'Marca não encontrada.'], 404);
+        }
+
+        $marca->delete();
+        return response()->json(['msg' => 'A marca foi removida com sucesso!'], 200);
     }
 }
